pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-bot-rag"
version = "0.1.0"
description = "RAG Chatbot with FastAPI, Chroma, LangChain, MLflow, and Optuna"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Marziye Askari" }
]

dependencies = [
    "langchain-openai",
    "langchain-chroma",
    "langchain-huggingface",
    "packaging",
    "opentelemetry-instrumentation-openai",
    "prometheus-client",
    "rouge-score",
    "optuna",
    "torch",
    "sentence-transformers",
    "mlflow-skinny",
    "uvicorn",
    "sqlalchemy",
    "langchain-classic",
    "langchain-community",
    "langchain-ollama",
    "langchain-core",
    "langchain-text-splitters",
    "langchain",
    "pyyaml",
    "pydantic-settings",
    "pydantic",
    "python-dotenv",
    "fastapi",
    "python-multipart",
    "psycopg[binary]",
    "pypdf",
    "boto3" ,
    "psycopg2-binary",
    "langchain_postgres",
    "botocore"
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "pytest"
]



[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "src*", "mlops*"]

[tool.black]
line-length = 100
target-version = ["py311"]
